+++ app/serializers/ad_hoc_for_authors_task_serializer.rb
@@ -1,0 +1,2 @@
+class AdHocForAuthorsTaskSerializer < AdHocTaskSerializer
+end