+++ spec/support/pages/page_fragment.rb
@@ -85,7 +85,7 @@ class PageFragment
   end
 
   def view_card_in_task_sidebar(card_name)
-    el = find('#paper-submission-tasks .task-disclosure-heading', text: card_name)
+    el = find('.task-disclosure-heading', text: card_name)
     el.click
     yield if block_given?
   end