+++ db/migrate/20160920175155_change_text_column_to_label.rb
@@ -1,0 +1,4 @@
+class ChangeTextColumnToLabel < DataMigration
+  RAKE_TASK_UP = 'data:migrate:adhoc_tasks:change_text_blocks_to_label'.freeze
+  # RAKE_TASK_DOWN = 'fill_me_in or delete me if unnecessary'
+end