+++ engines/tahi_standard_tasks/spec/mailers/tahi_standard_tasks/initial_decision_mailer_spec.rb
@@ -1,34 +1,30 @@
 require 'rails_helper'
 
 describe TahiStandardTasks::InitialDecisionMailer do
-
-  let(:paper) do
-    FactoryGirl.create(
-      :paper,
-      :with_integration_journal,
-      :with_creator,
-      title: 'Paper Title'
-    )
-  end
-
-  let(:task) do
-    FactoryGirl.create(:initial_decision_task,
-                       paper: paper,
-                       completed: true)
-  end
-
+  let(:paper) { FactoryGirl.build_stubbed(:paper, title: 'Paper Title') }
   let(:decision) do
-    paper.decisions.create!(
-      letter: "Body text of a Decision Letter",
-      verdict: "reject"
+    FactoryGirl.build_stubbed(
+      :decision,
+      letter: 'Body text of a Decision Letter',
+      verdict: 'reject',
+      paper: paper
     )
   end
+  let(:paper_creator) { FactoryGirl.build_stubbed(:user)}
 
   let(:email) do
     described_class.notify(decision_id: decision.id)
   end
 
   describe "#notify" do
+    before do
+      allow(Decision).to receive(:find).
+        with(decision.id).
+        and_return decision
+
+      allow(paper).to receive(:creator).and_return paper_creator
+    end
+
     it "sends email to the author's email" do
       expect(email.to).to eq([paper.creator.email])
     end