+++ CHANGELOG.md
@@ -8,8 +8,10 @@ guidelines from here: https://github.com/olivierlacan/keep-a-changelog
   allowing manuscripts to be published before all proofreading and copyediting is done.
 - Invitations can now be reordered through drag and drop. This also enforces
   rules automatically for which invitations can be reordered and into which groups
+- A user can connect their ORCID account with their author profile via the Authors cards on a manuscript
 - Display of article type in invitations
 - Caching of unsaved Discussion responses
+- Ad-hoc cards have editing and managing permissions
 - ORCID IDs are included in the metadata export to Apex
 ### Changed
 ### Deprecated
@@ -17,6 +19,7 @@ guidelines from here: https://github.com/olivierlacan/keep-a-changelog
 - Removed 'Assign Admin' card
 - Removed 'Available Task Types' section in journal administration
 ### Fixed
+- Emails are properly rendered in all clients
 ### Security
 
 ## [1.29.0] - {2016-11-03}